Move tunnel logic into separate module
